Prettier: split configs and use direct dependencies
