+++ package.json
@@ -35,8 +35,6 @@
     "eslint-plugin-unicorn": "^56.0.1",
     "eslint-plugin-vitest": "^0.5.4",
     "prettier": "^3.4.1",
-    "prettier-plugin-svelte": "^3.3.2",
-    "prettier-plugin-tailwindcss": "^0.6.9",
     "semver": "^7.6.3",
     "svelte": "^5.4.0",
     "tailwindcss": "3.4.15",
