+++ packages/prettier-config/package.json
@@ -4,29 +4,18 @@
     "access": "public",
     "provenance": true
   },
-  "version": "0.3.6",
-  "description": "Common Prettier configuration for Viam projects.",
-  "type": "commonjs",
+  "version": "1.0.0",
+  "description": "Viam's shared Prettier configuration.",
+  "type": "module",
   "files": [
     "**/*",
     "!tsconfig.json"
   ],
-  "types": "dist/base.d.cts",
-  "typesVersions": {
-    "*": {
-      "svelte": [
-        "dist/svelte.d.cts"
-      ]
-    }
-  },
+  "types": "./dist/prettier-config.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/base.d.cts",
-      "default": "./base.cjs"
-    },
-    "./svelte": {
-      "types": "./dist/svelte.d.cts",
-      "default": "./svelte.cjs"
+      "types": "./dist/prettier-config.d.ts",
+      "default": "./prettier-config.js"
     }
   },
   "repository": {
@@ -40,16 +29,6 @@
   ],
   "license": "Apache-2.0",
   "peerDependencies": {
-    "prettier": ">=3 <4",
-    "prettier-plugin-svelte": ">=3 <4",
-    "prettier-plugin-tailwindcss": ">=0.4.0 <0.7"
-  },
-  "peerDependenciesMeta": {
-    "prettier-plugin-svelte": {
-      "optional": true
-    },
-    "prettier-plugin-tailwindcss": {
-      "optional": true
-    }
+    "prettier": ">=3 <4"
   }
 }
